feat (#22) : Principal Details User 데이터 획득하도록 수정 되었음

+++ src/main/java/cinebox/service/ReviewService.java
@@ -3,10 +3,13 @@ package cinebox.service;
 import java.util.List;
 import java.util.stream.Collectors;
 
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 
 import cinebox.common.exception.movie.NotFoundMovieException;
 import cinebox.common.exception.review.NotFoundReviewException;
+import cinebox.common.exception.user.NoAuthorizedUserException;
 import cinebox.common.exception.user.NotFoundUserException;
 import cinebox.dto.request.ReviewRequest;
 import cinebox.dto.response.ReviewResponse;
@@ -18,6 +21,7 @@ import cinebox.repository.MovieRepository;
 import cinebox.repository.ReviewRepository;
 import cinebox.repository.UserRepository;
 import cinebox.security.JwtTokenProvider;
+import cinebox.security.PrincipalDetails;
 import jakarta.servlet.http.HttpServletRequest;
 import lombok.RequiredArgsConstructor;
 
@@ -50,25 +54,33 @@ public class ReviewService {
 		return reviewRepository.findById(reviewId).orElseThrow(()-> NotFoundReviewException.EXCEPTION);
 	}
 
-	public void updateReview(ReviewRequest reviewRequest, HttpServletRequest request) {
-		String token = jwtTokenProvider.getToken(request);
+	public void updateReview(ReviewRequest reviewRequest) {
+		Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+		PrincipalDetails userDetails = (PrincipalDetails) authentication.getPrincipal();
+		Long requestUserId = userDetails.getUser().getUserId();
 		Review review = selectReviewByReviewId(reviewRequest.getReviewId());
-		UserResponse reviewer = userService.getUserById(review.getUser().getUserId());
-		jwtTokenProvider.isUserMatchedWithToken(reviewer.getIdentifier(), token);
 		
-		Movie movie = movieRepository.findById(reviewRequest.getMovieId()).orElseThrow(()-> NotFoundMovieException.EXCEPTION);
-		User user = userRepository.findById(reviewRequest.getUserId()).orElseThrow(()-> NotFoundUserException.EXCEPTION);
-		
-    	Review updatedReview = Review.of(movie, user, reviewRequest);
-    	reviewRepository.save(updatedReview);
+		if(requestUserId.equals(review.getUser().getUserId())) {
+			Movie movie = movieRepository.findById(reviewRequest.getMovieId()).orElseThrow(()-> NotFoundMovieException.EXCEPTION);
+			User user = userRepository.findById(reviewRequest.getUserId()).orElseThrow(()-> NotFoundUserException.EXCEPTION);
+			
+	    	Review updatedReview = Review.of(movie, user, reviewRequest);
+	    	reviewRepository.save(updatedReview);	
+		} else {
+			throw NoAuthorizedUserException.EXCEPTION;
+		}
 	}
 	
-	public void deleteReview(Long reviewId, HttpServletRequest request) {
-		String token = jwtTokenProvider.getToken(request);
-		Review review = selectReviewByReviewId(reviewId);
-		UserResponse reviewer = userService.getUserById(review.getUser().getUserId());
-		jwtTokenProvider.isUserMatchedWithToken(reviewer.getIdentifier(), token);
+	public void deleteReview(Long reviewId) {
+		Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+		PrincipalDetails userDetails = (PrincipalDetails) authentication.getPrincipal();
+		Long requestUserId = userDetails.getUser().getUserId();
+		Long reviewerId = selectReviewByReviewId(reviewId).getUser().getUserId();
 		
-		reviewRepository.deleteById(reviewId);
+        if (requestUserId.equals(reviewerId)) {
+    		reviewRepository.deleteById(reviewId);
+        } else {
+        	throw NoAuthorizedUserException.EXCEPTION;
+        }
 	}
 }
